# api/open-api/swagger.yaml
openapi: 3.1.0
info:
  title: Swagger - OpenAPI 3.1
  description: API documentation for the final project.
  # termsOfService: http://swagger.io/terms/
  # contact:
  #   email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
# servers:
#   - url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /api/users:
    get:
      tags:
        - users
      summary: List all users
      description: This can only be done by the logged in user
      operationId: listUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/{userId}:
    get:
      tags:
        - users
      summary: Retrieve details of a specific user
      description: This can only be done by the logged in user
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: The ID of the user to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: User not found
    patch:
      tags:
        - users
      summary: Update details of a specific user
      description: This can only be done by the logged in user
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: The ID of the user to modify
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: User not found
    delete:
      tags:
        - users
      summary: Delete user
      description: This can only be done by the logged in user
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: The name of the user that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: User not found
  /api/users/{userId}/sync:
    post:
      tags:
        - users
      summary: Sync user listings
      description: Sync user listings
      operationId: syncUserListings
      parameters:
        - name: userId
          in: path
          description: The ID of the user to sync listings
          required: true
          schema:
            type: string
        - name: provider
          in: query
          description: The provider to sync listings from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: User not found
  /api/providers:
    post:
      tags:
        - providers
      summary: Create a new provider
      description: This can only be done by the logged in user
      operationId: createProvider
      requestBody:
        description: Create provider object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Provider already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - providers
      summary: Delete provider
      description: This can only be done by the logged in user
      operationId: deleteProvider
      requestBody:
        description: Delete provider object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request
        '404':
          description: Provider not found
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [10]
        username:
          type: string
          examples: [theUser]
        firstName:
          type: string
          examples: [John]
        lastName:
          type: string
          examples: [James]
        email:
          type: string
          examples: [john@email.com]
        password:
          type: string
          examples: ['12345']
        phone:
          type: string
          examples: ['12345']
        userStatus:
          type: integer
          description: User Status
          format: int32
          examples: [1]
    Provider:
      type: object
      properties:
        provider:
          type: string
          examples: [1234567890]
        userId:
          type: string
          examples: [1234567890]
...
    ApiResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header